move(slack): command 패키지와 result 패키지를 dto 패키지로 통합
