+++ slack/slack/src/main/java/com/luckylogistics/slack/SlackApplication.java
@@ -2,11 +2,12 @@ package com.luckylogistics.slack;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
 import org.springframework.cloud.openfeign.EnableFeignClients;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
 @SpringBootApplication
-// @EnableDiscoveryClient
+@EnableDiscoveryClient
 @EnableFeignClients
 @EnableJpaAuditing
 public class SlackApplication {
