+++ slack/slack/src/main/java/com/luckylogistics/slack/application/dto/SlackEmailCheckResult.java
@@ -1,4 +1,4 @@
-package com.luckylogistics.slack.application.result;
+package com.luckylogistics.slack.application.dto;
 
 import lombok.Builder;
