+++ slack/slack/src/main/java/com/luckylogistics/slack/application/external/AiServiceClient.java
@@ -2,8 +2,8 @@ package com.luckylogistics.slack.application.external;
 
 import java.time.LocalTime;
 
-import com.luckylogistics.slack.application.result.AiPromptCreatedResult;
-import com.luckylogistics.slack.application.result.OrderCreatedResult;
+import com.luckylogistics.slack.application.dto.AiPromptCreatedResult;
+import com.luckylogistics.slack.application.dto.OrderCreatedResult;
 
 public interface AiServiceClient {
