+++ slack/slack/src/main/java/com/luckylogistics/slack/application/service/SlackService.java
@@ -11,18 +11,17 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.SerializationFeature;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
-import com.luckylogistics.slack.application.command.EmailCheckCommand;
-import com.luckylogistics.slack.application.command.StatusUpdateCommand;
+import com.luckylogistics.slack.application.dto.EmailCheckCommand;
+import com.luckylogistics.slack.application.dto.StatusUpdateCommand;
 import com.luckylogistics.slack.application.event.SlackKafkaEventPublisher;
 import com.luckylogistics.slack.application.external.AiServiceClient;
 import com.luckylogistics.slack.application.external.SlackClient;
-import com.luckylogistics.slack.application.result.AiPromptCreatedResult;
-import com.luckylogistics.slack.application.result.OrderCreatedResult;
-import com.luckylogistics.slack.application.result.SlackEmailCheckResult;
-import com.luckylogistics.slack.application.result.SlackMessageResult;
+import com.luckylogistics.slack.application.dto.AiPromptCreatedResult;
+import com.luckylogistics.slack.application.dto.OrderCreatedResult;
+import com.luckylogistics.slack.application.dto.SlackEmailCheckResult;
+import com.luckylogistics.slack.application.dto.SlackMessageResult;
 import com.luckylogistics.slack.domain.entity.SlackMessage;
 import com.luckylogistics.slack.domain.repository.SlackRepository;
-import com.luckylogistics.slack.domain.vo.Status;
 import com.luckylogistics.slack.infrastructure.external.kafka.event.OrderCreatedEvent;
 
 import lombok.RequiredArgsConstructor;
